Scene Changing & Build Storage Fix
We've implemented the ability to swap between scenes once all players are ready, going from the Mech Building to a Gameplay scene (Currently just loads the appropriate models)

Additionally, because our ServerManager isn't syncing the ClientData with the clients (Which we shouldn't as it's a lot of data to be sending over), we've changed how we're using it so that we can send over the indicies of the frame, leg, weapons, and abilities instead.
This lets clients load the appropriate models and build data for all the players in the newly loaded scenes.