Basic Server Authoritative Movement
We've implemented a basic Server Authoritative Movement to the players in the Gameplay scene.

Our input is being received client-side, and then passed to the server. We're currently doing the server passing through directly calling our PlayerMovement script's ServerRpc from input, but we're going to want to handle this differently so that the ClientSideInput isn't dependant on other scripts.