ActionConfig Headers
We've added in headers to the ActionConfig class to make it easier to read in the inspector.

We were trying to make a custom inspector for it so that only the neccessary sections showed depending on the selected ActionLogic type of 'Logic', but we ended up stopping as we didn't have a clear enough view of how we wanted to structure it.