ActionConfig replaced with ActionDefinitions
We've replaced ActionConfig with ActionDefinition (The script still exists currently), with ActionDefinition letting us better specify sections of actions, and through ActionEffects it also lets us only have one type of base 'Action', with the definitions using ActionEffects to construct their output.

This does make it a bit more complex for simple actions (Something that can be mitigated a little with structuring the editor), but it makes the system more modular, expandable, and simpler to make complex actions.

To-Do: Look at whether we should have the ActionDefinitions be ScriptableObjects and Action just be a class (Currently couldn't do because of how we're creating and pooling Actions)