RaycastTargeting & TargetableTypes IsValidTarget
We've implemented RaycastTargeting as a type of ActionTargeting, along with implementing the 'IsValidTarget' function for the TargetableTypes enum.

Additionally, we've made it so that we only run our targeting logic if we have a bool telling us that we are using effects of that type. (To-do: Automate this process rather than having it as a serialized bool that we need to tick, but still keep it performant (Not checking length every call)).