Serializing Origin Transform & Action Timings synced from Server,
We've made it so that we can serialize the Origin Transform by having it be a ulong Network Object id (Using the root NetworkObject of the weapon). This lets us use it on the clients for visualisations.

Additionally, we've made the Action triggering be dependant on the server time rather than local time, which should make it more synchronised (Hopefully. If not, we can change back to Time.time)