ActionEffects Start
We've started work on ActionEffects, which will be used by some actions (E.g. Weapon Attacks) in order to perform their effects.

We are aiming to have ActionEffects for dealing damage, healing, adding/removing buffs, and spawning objects (E.g. Mines)

We've still to do the implementation for most of the ActionEffects