SpawnObjectEffect Work & HitForward Calculation Fix
We've worked on the SpawnObjectEffect, getting it to a point where we can calculate the spawn positions and spawn our objects via an object pool.
Note: We currently aren't handling max batches correctly, and an infinite number of objects can be spawned.

We've also fixed our calculation of HitForward which would occasionally return the wrong value when the vectors were ever-so-slightly off (Typically at object edges)