Projectiles Start
We've started work on projectiles, implementing regular and explosive projectiles.

We also made an attempt on Seeking Projectiles, but never got anything we were happy with. Our attempts included:
- Using rigidbodies and accelerations
- Using rigidbodies and overriding the velocity
- Directly modifying the position, rotating towards the target, moving at a fixed speed
- Directly modifying the position, rotating towards the target, and using acceleration
- Directly modifying the position and using a PID (We figured it's too complex for our game though)

We also had methods of getting the target position ranging from a Transform target to a Func<Vector3>. If we cache the origin & desired direction (Or target transform), then we found that the Func<Vector3> is the best option for flexability (Allowing for seeking to a transform or a raycast from the character)

To-do:
- Seeking Projectiles (For missiles)
- Activate Action Effects (Muzzle Flashes)