Action Visuals
We've started work on Action visuals, including porting our spawning of a SpecialFXGraphic to this system (Which we've also now made pooled).

We can have these visuals apply during start, update, end, or cancel based on the value of a flags enum, and are triggered via the OnClientX functions.
As a part of this, we've also iterated on the ClientActionPlayer's OnUpdate() and various cancel functions to put them in line with the ServerActionPlayer's corresponding functions.