SpawnObjectEffect Parenting/Attachment Improvement & Pooling Fixed
We've worked on the parenting of the objects spawned via the SpawnObjectEffect, as our previous method would lead to issues with parenting to NetworkObjects.
Now, we're simulating parenting by having the object set its position & rotation based on the initial local position, forward, and up vectors. This circumvents the NetworkObject parenting issues.

We've also fixed several errors with the returning of the SpawnableObjects to the pool, most of which were caused by the parenting or retrieving an active instance and not clearing event subscriptions.

Note: SpawnableObjcet Attaching is not synced on clients currently. We'll likely to this by estimation, similar to how projectiles are currently, as NetworkTransforms would consume too much bandwidth when their numbers start to increase