Charging UI Gradual Change & Multi-Weapon Support
We've implemented the ability to have the Charging UI gradually change when we fire a shot (With this working with bursts).
We've also made the UI work with multiple actions via the WeaponSlotIndex (Which we're soon going to change to AbilitySlotIndex). (Note: Actions are still referencing the UI. Ideally, this should be inverted)

Additionally, we made a basic weapon UI example for the primary weapon in the GameplayTestScene