UI Separation from Action Logic & Multiple UI Support
We've separated the calling of the UI logic from the Action logic by having Action call events which the UI subscribes to. We've also made it so that this is supported with multiple clients.

Additionally, we've made it so that we can have multiple UI elements for the same slot identifier.