Customisation UI Work
We've worked on the Customisation UI, getting to a point that we are happy with starting implementation on.

The UI allows players to select their desired frame from a menu accessed via button press, their attachment point choices via a menu with tabs for each slot (Located on the left of the screen), view their currently selected option's name, description, effects, & stats, and their currently selected frame's stats.