Frame Selection Menu Opening & Tab Selection
We've added the ability to open the Frame Selection Menu and have made it so that we can move between tabs in the Attached Slottable Selection and Frame Selection menus with input.

Note: Currently we can still navigate the UI while in the Frame Selection Menu. Maybe separate input groups?