Customisation Changeable in Game Start
We've started working towards having Player Customisation be something that is changeable in-game, as opposed to just in the lobby.

We're going to achieve it by letting the player toggle aspects of their build while respawning, giving them access to a menu similar to the main menu.

We've so far made it so that finalising a build no longer deletes the unused GameObjects, and have started making it so that in-game players can have their build changed and their graphics and actions will update.

Next Steps:
- Clean-up Customisation Finalisation Code (Removing it?)
- Address the duplication of calls when a player is spawned (PlayerManager and PlayerCustomisationManager)
- (Can leave till after merge) Implement a way to open the Player Customisation Screen during
- Finish Actions for Stealth & Spawning (Allows us to finally merge this branch back into Main) (Stealth requires us to think about Buffs)