Stealth Action & StatusEffects Start
We've started implementing Status Effects to let us create the Stealth Action (Which we've done). Along with this, we started work on tracking Heat.

StatusEffects can currently have a lifetime, a retrigger delay, generate or dissapate heat from the owner each frame, be considered a Buff or Debuff, and have a Stacking Type set, as follows:
- Retrigger: Reset the lifetime of the existing effect and re-trigger its 'OnStart' method.
- Reset Duration: Reset the lifetime of the existing effect
- Add Duration: Add the lifetime of the effect to the existing effect
- In Parallel: Run this effect in parallel with others of the same type
- Toggle: Cancel the active effect and don't restart it

To Look Into:
- Actions that persist until you trigger them again, at which point they cancel. We can then use this rather than having status effects affect the Heat of the player