Action Cooldown Visuals Fix
We've fixed the visuals not conforming to Action Cooldowns and similar things by having the actions trigger their visuals when they are started, rather than when the action is requested.
Note: For client anticipation we should change this so that an anticipation function is called, we check using local values, and then display the visuals.

We're wanting to add the ability to have actions that are on cooldown be queued with input (Or continually trigger input) so that you don't need to re-activate in order to trigger them.