Player Health Error Fix
We've fixed the error with the Debug UI reading the player's max health before their build is set by now using a NetworkVariable for the BuildData (Which should also make it easier for modifying in the future). We've also added initialisation functions to the OnChange event for Health & Heat to keep them updated.