Player Camera Start
We've started work on the Player's camera (Currently the Third Person view), implementing rotation for the player and fixing movement to work with it. We're also pointing the graphics towards where the player is looking, rather than the camera's facing direction, to account for camera offsets on the local x.

We've also installed the Cinemachine package.