Swallowtail Model Test
We've created a test for the Medium Sized Frame based on the SSC Swallowtail from the TTRPG Lancer.

We are having some issue with the model, and have temporarily inverted the Primary and Secondary triggers from what they're named due to the sides they appear on.

Our plan is that the main chassis of the model will tilt to handle vertical offsets, however we haven't yet implemented this.

Frames:
- Large: Polygon Mech (Tilt at waist)
- Medium: Custom based on Swallowtail
- Small: Custom based on Swallowtail w/ Central Weapon?