Crosshair Offset Fix
We've fixed a bug with the Targeting Crosshairs not appearing at the correct position in the screen when there was no obstruction.
This was caused by the fact that they weren't aligned to the camera when determining their position, but we're now using a plane created by the camera to keep the estimated positions aligned no matter the offset or horizontal rotation difference.