Vertical Rotation & Target Rotation Fix
We've implemented vertical rotation for the player with a separate pivot than the horizontal rotation (Vertical Rot Pivot needs to be a child of the Horizontal). We also have an offset for this so that we can align our rotation to our weapon positions better (We could automate this, but manual gives us more control and is easier).

Additionally, we've fixed our calculation for the crosshair world target position & player target rotation, making our crosshair more reliable.

To-do:
- Action FX (And possible Logic too) isn't accounting for rotations again. We should figure out why.
- The crosshairs currently treat all actions as if they will raycast. Add in options to specify the crosshair & to disable it from actions (For Projectiles & Self-Targeting/Non-Raycast AoEs respectively)
- Add in colour changing to the crosshairs based on if they are targeting nothing, are targeting a friendly, or are targeting an enemy. Also add in a alpha/value change if they are obstructed.