IActionSource & Actions Affected by Rotation again
We've implemented an 'IActionSource' interface to allow us to pass references to Action origin components (E.g. Origin Transform) via a NetworkObjectId.

This has allowed us to have actions be affected by updating rotation throughout their execution (If we pass the ID, otherwise Position and Direction are passed and used as usual).

We've also removed the GFXSection class as it only had a single subclass and no direct references, and have fixed any issues we've found our changes to have created.