Rotation Offset Auto Calculation
We've made it so that we're automatically calculating the vertical rotation offset rather than it being a preset value, which lets us account for differing heights of weapons.

Note: Currently this means that for mechs such as the Swallowtail, having a LMG/Railgun and a Missile Launcher will cause both to be slightly vertically offset.