Basic Respawning
We've implemented basic respawning, with player death preventing input and respawning re-enabling input and teleporting the player to a desired position.

Subsequently, we've also implemented notifying the owning client when their player dies. This is currently being used to disable and re-enable input, however we will also use it for showing the Respawning Screen and similar things.