Player Target Rotation Bugfix Start
We've started work on fixing a bug with our calculation of the player's target rotation which is causing them to incorrectly aim at the target position.

We're getting close to this by using TransformDireciton with the Camera on the vertical pivot offset, however this is slightly off for nearby targets due to the horizontal offset.

To-do:
- Finish solving this issue (Reference CameraControllerTest.cs ~Line 178)
- Ensure that resetting of rotation works properly during respawning