Lobby Work & Player Name Bugfix
We've worked some more on the Lobby, getting it to where we can perform the game start (No scene change implementation yet) and change the selected Slottables, with data synced through the NetworkBuildData on the PersistentPlayer.

We've also fixed a error with our implementation of the Player Name, which was causing us to use the PlayerId rather than their name and caused an error in the FixedString32Bytes conversion, breaking a few scripts.