Frame Changing Re-implementation
We've re-implemented Frame Changing in the lobby.

We've also removed the PlayerCustomisationManager as it is no longer going to be in use, instead replaced by the NetworkBuildState on the PersistentPlayer (Note: We haven't yet made the Build Changing server-side. It is being changed on the clients and relayed to the server through its NetworkVariable)