Health/Life Work & Build in Gameplay
We've worked on the Health/Life scripts and how the builds interact with gameplay.

We are currently using a class that can hold a reference synced across clients in order to have the ServerCharacter receive changes to the build, but are somewhat unhappy with this.

Additionally, we've removed the Health UI for now