In-Game Builds Work
We've worked on how we're handling in-game builds, changing from a reference to instead having a BuildData variable in the ServerCharacter which can be populated as desired. We are then populating this from the Player class on its Network Spawn using values from the PersistentPlayer, and are listening to changes from the PersistentPlayer to update appropriately.

Additionally, we're now updating the BuildData reference within the NetworkBuildState whenever we next retrieve it after a change (And the first time we ever retrieve it), ensure that our data is always up-to-date. This also makes it so that joining players see the correct data when they join (However they currently cannot perform input for some reason)

We've also added a 'ReadOnly' attribute, letting us view items in the inspector without being able to edit them, and have removed the GameManager reference from the RespawnScreenUI as we no longer have a GameManager (It's getting replaced by the GameState) and was causing errors upon player death.