Ordered Mid-Game PlayerData, Leaderboard & PersistentPlayer Saving Fix
We've implemented sorting of our PlayerData on clients (Using int indicies rather than duplicating the structs) so that the Leaderboard can easily display the player data in order of their Score.
For mid-game connections or reconnecting clients, we're initialising the sorted array if the NetworkList was initialised before our OnListChanged subscription (So otherwise we wouldn't receive that initial data until it was changed)

We are currently using a BubbleSort algorithm for fully sorting the array, and are sorting upwards using direct comparison and swapping when we insert/change a value

Additionally, we've fixed a bug in our PersistentPlayer which occured due to us calling base.OnDestroy() before RemovePersistentPlayer(), leading to our NetworkVariables being disposed before we tried to access them.