Late-Join Initial Build Prompting
We've implemented a quick method of having late-joining players need to choose their build.

As such, we've also removed the Mid-Game Customisation UI's requirement of having the 'Player' class be spawned, instead using the static 'PersistentPlayer' OnLocalPlayerBuildChanged event.
We've also renamed the file to MidGameCustomisationUI to match the name of the class found within.