CameraControllerTest Raycast Fix
We've fixed a bug in CameraControllerTest script which was causing us to have an incorrect target position value. This was due to RaycastNonAlloc not returning values in order, so we've changed to use RaycastAll, ordering it by ascending distance, and then taking the closest 2 values and taking the non-player one.