Projectile Hit Effect Syncing
We've fixed a bug with the Projectile's not syncing on clients other than their triggering one due to us missing a function call.

We've also removed the OnCollisionEnter function from the Projectile as we previously changed how their collisions are being calculated to raycasts so that they were no longer phasing through some walls.