Teleporation Breaking Rotation & Client Visuals Not Appearing Bugfixes
We've fixed the bug causing player rotation & movement to break when they are respawned. This was due to us incorrectly setting the root transform's rotation. We are now doing a TryGetComponent check in ServerCharacterMovement (Would replace if we had more time) to set the desired rotation.

We've also fixed a bug where the client visuals spawned by the HitEffectManager would not include the triggering client when triggered on a non-host client (Changed the 'Self' function to 'TriggeringClient').