Cheats Instance Check & Smallcaps Disable
We've implemented an instance check for the DebugCheatManager in the ChatManager so that it only attempts to process a cheat if the DebugCheatManager exists (During Gameplay), preventing errors from doing so, and we notify the triggering client if this occurs.

We also disabled the Smallcaps option for the chat to make it easier to differentiate between Uppercase and Lowercase letters.