More Code Comments & Lobby Voting Initial Display Fix
We've added more code comments to the project & have fixed a bug with the Lobby Voting display where elements weren't showing the current votes if the player entered the leaderboard for the first time after a change had occured.

Now, we perform a forced update when we receive our dependancy and when the button is enabled to ensure that the value displayed is always correct.