Comments & CustomisationDummyManager Bugfix
We've added some more comments & have fixed a bug with the CustomisationDummyManager which was causing it to incorrectly determine that we didn't want to spawn player dummies in the lobby.

We also changed the Exception thrown when there are too many players in the lobby to instead be a LogWarning & early return, allowing for more players to join without breaking the game.